requirements.txt
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
